Create packages, optimize imports
